CL-Project/python-microservice/requirements.txt
Flask
faster-whisper
google-generativeai
networkx
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio